dates.c: Revert to 'return -1' for first_dayofweek_of_{year,month}()
This program needs to parse many events that beyond the specified date
range, which can cause 'first_dayofweek_of_{year,month}()' to fail.
However, this is not expected.  So revert to 'return -1' instead of
abort the program and check the return value in the callers.

This fix commit bbaffe6c1f8c919178414fc1419e983b1fed8f29.