Make '-t' option use a date string like date(1)
Rewrite the Mktime() function to parse a date string of format
'[[[cc]yy]mm]dd', similar to the format used by date(1).

While there, fix the DEBUG printf in Mktime().  The behavior is
undefined whether 'asctime(&tm)' would be called after or at the same
time as the 'mktime(&tm)' in the function parameters.

Credit:
https://github.com/openbsd/src/commit/c60f2790b51cb296bde66867b870b4fb82e485e6