calendar: Remove '-l' option (deprecate by '-L' option)
The '-l' option that specifies the longitude is only used to set the
timezone, which can be also specified by the '-U' option.

Furthermore, the new '-L' option can be used to specify the longitude
as well as latitude and elevation.  Thus remove the '-l' option.

In addition, remove the unused 'EastLongitude' member from
'struct cal_options'.