parsedata.c: Copy extradata to avoid overwrite by event repeation
The 'extradata' field is currently used by solar and lunar events to
record the exact time of the events.  If the event (e.g., NewMoon)
repeats in the given date range, the extradata of each such event would
overwrite.  Avoid such a overwrite by copying the extradata.

This reverts commit e0741123d61dbad78e0dfdf4b282af0db0039902 that
introduced such an issue.