io.c: Implement cal_readentry() to better read calendar files
The cal_readentry() function provides a much simpler interface that will
be used by cal_parse() to read calendar files.  It can handle the case
of that an event can have multiple content lines that start with tabs.

By using the new cal_readentry() function, cal_parse() need not to track
the continuation lines of the previous event, and thus can be greatly
simplified.  This refactor will be followed.