calendar.primety: Remove the CALENDAR reset in the end
This calendar was imported from OpenBSD, whose calendar(1) supports
to set 'CALENDAR=Julian' to switch to the Julian calendar, and it
requires to reset the calendar switching at the end of the calendar file
by setting 'CALENDAR='.

I don't want to follow the above way and plan to reset the calendar to
the default Gregorian at the end of the calendar file automatically,
similar to the way of dealing with 'LANG' variable.

While there, captialize the 'julian' to 'Julian' for consistency.