parsedata.c: Check for month numbers before month names
In some cases, such as the 'zh_CN.UTF-8' locale on macOS, the short
national month name can be the *same* as the month number.  This
confuses the date parsing to the wrong branch and cause the parsing to
fail.

Check the month numbers *before* month names to fix the above issue.
Note that the 'is_onlydigits()' has already been fxied to guarantee
the string cannot be a single star '*'.  So this reorder won't cause
issues to parse date of style '<month> *'.

Reported-by: Yongkai ZHU