calendar: Allow '~/.calendar' to be absent
We already allow the absence of user's calendar file
(~/.calendar/calendar) and let calendar(1) fallback to the system
default (/etc/calendar/default).

However, it was having an issue that the '~/.calendar' must exist.  This
fixes this issue and allows that directory to be also absent.  In that
case, calendar(1) fallbacks to the system calendar directory
(/etc/calendar).

Reported-by: Yongkai ZHU